backend/requirements.txt
# Core Dependencies
paho-mqtt>=2.0.0
python-telegram-bot>=20.0

# Web Framework
Django>=5.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Database
sqlalchemy>=2.0.0

# Data Processing
openpyxl>=3.1.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization (for future web dashboard)
plotly>=5.15.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
